requirements.txt
pyfstab~=0.2.0
python-dotenv~=1.0.1
tabulate~=0.9.0